Support for xrefs between nested prototypes
Now field cross references ($somefield) can be done both from parent
prototype to a child field, and from a nested prototype to a parent
field.

prototype parent
{
	num		[0 10]
	sum		add $childA.val $childB.val
	working	or $done $started
	childA	{
		val	[0 10]
		done	bool
	}
	childB	{
		val	[0 10]
		started	bool
	}
}

In this case a reference to val needs to be qualified with the containing prototype, such as $childA.val or $childB.val, but $done and started are adequate reference regardless of from where the reference is made (from childA, childB or the parent), as they both uniquely identifies a field.