Added macro-function capability to the DSL
Now a field-definition (generator) can be piped into a macro function
that has access to the generator definition, and the evaluator, and thus
can choose how to evaluate the generator.

A field definition such as;

    lotto   int [0 100] | repeat [5 10] | sort

will cause (the macro) sort to be called, with "repeat [5 10]" as the
inner generator. The sort macro simply evaluates the inner generator,
expecting a list of elements, and returns the sorted list. In this case
sort could have been a non-macro function, however everything after a
pipe is considered a macro function at this point.

The repeat-macro generates a repetition count using its first parameter,
and then uses it's inner generator to generate that many values,
returning a list of the generated values.

As repeat is called with "int [0 100]" as the inner generator, it will
generate a list of 5-10 integer numbers in the range 0-100.