Separated DSL-macros from pipelining
Now pipelining is simply transforming

func a | func b

into

func b (func a)

Whereas macros are ordinary functions in the dslfunctions namespace,
annotated with {:fabmacro true}. The parameters for these functions are
not evaluated before the function is called, instead the function
receives its own field definition, and are able to introspect on the
definition of its parameters.

A macro is responsible for using its definition to generate a value,
possibly manipulating the definition.  Repeat is now a macro function
that generates a number of values by repeatedly using the definition  of
its last parameter, to generate a value. The first parameter is used
once to generate a value, which then controls the number of repetitions.